Made project structure more readable, cleaned up dependencies.
