+++ backend/src/backend/utils/utils.py
@@ -1,0 +1,11 @@
+import json
+
+
+def parse_message(message: str) -> dict:
+    # The message is a json string, turn it into a dictionary
+    try:
+        message_dict = json.loads(message)
+    except json.JSONDecodeError:
+        print("Error parsing message")
+        return {"error": "Invalid JSON format"}
+    return message_dict
